Refactored WikiPDFExporter in many classes
Also cleaned up using
