+++ AzureDevOps.WikiPDFExport/ILogger.cs
@@ -6,4 +6,9 @@ namespace azuredevops_export_wiki
     {
         void Log(string msg, LogLevel logLevel = LogLevel.Information, int indent = 0);
     }
+
+    public interface ILoggerExtended : ILogger
+    {
+        void LogMeasure(string msg);
+    }
 }
