+++ AzureDevOps.WikiPDFExport/IWikiDirectoryScanner.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace azuredevops_export_wiki
+{
+    internal interface IWikiDirectoryScanner
+    {
+        IList<MarkdownFile> Scan();
+    }
+}
