+++ AzureDevOps.WikiPDFExport/IWikiPDFExporter.cs
@@ -2,6 +2,6 @@ namespace azuredevops_export_wiki
 {
     public interface IWikiPDFExporter
     {
-        
+
     }
 }
