+++ AzureDevOps.WikiPDFExport/MarkdigExtensions/DeepLink.cs
@@ -1,13 +1,13 @@
 using Markdig;
+using Markdig.Helpers;
 using Markdig.Parsers;
-using Markdig.Syntax;
 using Markdig.Renderers;
 using Markdig.Renderers.Html;
-using System.Collections.Generic;
-using System.IO;
-using Markdig.Helpers;
+using Markdig.Syntax;
 using Markdig.Syntax.Inlines;
 using System;
+using System.Collections.Generic;
+using System.IO;
 
 
 namespace azuredevops_export_wiki
